Playground/Project-2-T1-2026-Hashem_Ved_Areen/Hashem/smart_street_lighting/requirements.txt
# RAG Pipeline
llama-index>=0.12.0
llama-index-vector-stores-chroma>=0.4.0
llama-index-embeddings-openai>=0.3.0
llama-index-llms-anthropic>=0.6.0
llama-index-llms-openai>=0.3.0
llama-index-readers-file>=0.4.0
chromadb>=0.5.0

# LLMs
anthropic>=0.40.0
openai>=1.50.0

# Data processing
pandas>=2.0.0
pydantic>=2.0.0
numpy>=1.24.0

# Spatial analysis
scikit-learn>=1.3.0
folium>=0.15.0

# PDF parsing
pymupdf>=1.24.0

# Visualization (used by demo notebook)
matplotlib>=3.7.0

# Database
psycopg2-binary>=2.9.0

# Web framework
fastapi>=0.100.0
uvicorn>=0.20.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
